Making Windows happy

+++ src/goop.cpp
@@ -10,10 +10,10 @@ Goop::Goop(MeshID mesh_id, ShaderID shader_id, glm::fvec3 pos)
 
     static std::default_random_engine e;
     static std::uniform_int_distribution<> dis(0, 359);
-    this->rotation_angle = dis(e);
+    this->rotation_angle = (float)dis(e);
 
     static std::uniform_real_distribution<> dirdis(-1.0f, 1.0f);
-    this->rotation_dir = dirdis(e);
+    this->rotation_dir = (float)dirdis(e);
 
     this->render_obj.mesh = mesh_id;
     this->render_obj.shader = shader_id;
